backend/requirements.txt
google-genai
python-dotenv
jsonschema
requests
newspaper3k
lxml_html_clean
pydantic 
aiohttp 